[FIX] Unit tests cases

+++ MicroserviceWarehouse/product/src/main/java/com/olvera/warehouse/product/controller/ProductController.java
@@ -77,7 +77,7 @@ public class ProductController {
     })
     @GetMapping("/usersProducts")
     public ResponseEntity<PageResponse> getUsersProducts(
-            @RequestParam(value = "userId", defaultValue = "1", required = true)Integer userId,
+            @RequestParam(value = "userId", defaultValue = "1", required = true)Long userId,
             @RequestParam(value = "pageNo", defaultValue = DEFAULT_PAGE_NUMBER, required = false) int pageNo,
             @RequestParam(value = "pageSize", defaultValue = DEFAULT_PAGE_SIZE, required = false) int pageSize
     ) {
